Ensure from_box works if boxed.addr() == 1.
SmallBox::from_box will reuse the allocation from its argument as
ptr in smallbox. However, said allocation might be equal to 1,
which is the INLINE_SENTINAL value.

This would cause the resulting smallbox to read from its
uninitialized inline storage space when dereferenced.

Note that this problem likely cannot be triggered on a desktop OS
since the first page is typically set up to segfault when read.
Thus if the type is not a ZST, then the allocator won't have given
out a pointer in the first page, and if the type is a ZST,
then reading from the uninitialized storge space is perfectly fine
behavior.

However, bare metal programs could concivibly run into this
problem, so this commit causes from_box to copy from boxed into
a newly created smallbox if boxed.addr() == INLINE_SENTINEL.